v0.0.16 -- Fixed a bug that would allow users to make unconvential reservations
Basically, prior to this version, if a student selected a slot with a TA (say
TA #1) and the student selected the "submit reservation" option as the action to
perform, all of this while the slot they were interested in was already taken,
then no reservation would be made. But the final option selection would still
be "submit reservation". So if the student, then switched TAs (without changing
the day of the week) to another TA (say TA # 2), it would then submit the
reservation without checking if TA # 2 had a valid slot that day.

Now with version v0.0.16, if the student tries to make an invalid reservation,
the action to perform selector would be reset. So when the student switched TAs,
no reservations would be automatically created. The student would have to select
the "submit reservation" option a second time, and if the check did not pass
the reservation would not be made.